+++ config/default.yaml
@@ -15,7 +15,7 @@ services:
     systemUsername: 'dummy_user'
     systemPassword: 'dummy_password'
   case:
-    url: 'http://ccd-data-store-api-adoption-cos-api-pr-249.service.core-compute-preview.internal'
+    url: 'http://ccd-data-store-api-adoption-cos-api-pr-259.service.core-compute-preview.internal'
   postcodeLookup:
     url: 'https://api.os.uk/search/places/v1'
     token: 'NEED TO INSERT SECRET'
