+++ src/main/steps/review-pay-submit/check-your-answers/content.test.ts
@@ -30,8 +30,8 @@ const enContent = {
   },
   sectionTitles: {
     applicationDetails: 'Application details',
-    adoptionagencyOrLA: 'Adoption agency or local authority details',
-    additionalAoptionagencyOrLA: 'Additional adoption agency or local authority details',
+    adoptionagencyOrLA: 'Local authority details',
+    additionalAoptionagencyOrLA: 'Adoption agency or local authority details',
     socialWorkerDetails: "Child's social worker details",
     applicantDetails: "Applicant's details",
     firstApplicantDetails: "First applicant's details",
@@ -53,7 +53,7 @@ const enContent = {
     emailAddress: 'Email address',
     nameOfContact: 'Name of contact',
     emailOfContact: 'Email address of contact',
-    teamEmailAddress: 'Team email address',
+    childLocalAuthority: "Child's local authority",
     additionalAdoptionAgency: 'Additional adoption agency',
     fullName: 'Full name',
     previousNames: 'Previous names',
@@ -73,8 +73,8 @@ const enContent = {
     court: 'Court',
     date: 'Date',
     courtOrder: 'Court order',
-    siblingCourtOrders: 'Sibling court orders',
     siblingOrHalfSibling: 'Child siblings or half siblings',
+    siblingCourtOrders: 'Sibling court orders',
     siblingRelation: 'Sibling relation',
     familyCourtName: 'Family court name',
     uploadedDocuments: 'Uploaded documents',
@@ -112,8 +112,8 @@ const cyContent = {
   },
   sectionTitles: {
     applicationDetails: 'Manylion y cais',
-    adoptionagencyOrLA: 'Manylion yr asiantaeth fabwysiadu neu’r awdurdod lleol',
-    additionalAoptionagencyOrLA: 'Manylion ychwanegol yr asiantaeth fabwysiadu neu’r awdurdod lleol',
+    adoptionagencyOrLA: 'Local authority details (in welsh)',
+    additionalAoptionagencyOrLA: 'Adoption agency or local authority details (in welsh)',
     socialWorkerDetails: 'Manylion gweithiwr cymdeithasol y plentyn',
     applicantDetails: 'Manylion y ceisydd',
     firstApplicantDetails: 'Manylion y ceisydd cyntaf',
@@ -135,7 +135,7 @@ const cyContent = {
     emailAddress: 'Cyfeiriad e-bost',
     nameOfContact: 'Enw’r unigolyn cyswllt',
     emailOfContact: 'Cyfeiriad e-bost yr unigolyn cyswllt',
-    teamEmailAddress: 'Cyfeiriad e-bost y tîm',
+    childLocalAuthority: "Child's local authority (in welsh)",
     additionalAdoptionAgency: 'Asiantaeth fabwysiadu ychwanegol',
     fullName: 'Enw llawn',
     previousNames: 'Enwau blaenorol',
