+++ src/main/steps/review-pay-submit/check-your-answers/content.ts
@@ -12,6 +12,7 @@ import {
   childrenPlacementOrderSummaryList,
   childrenSummaryList,
   familyCourtSummaryList,
+  localAuthoritySummaryList,
   otherParentSummaryList,
   siblingCourtOrderSummaryList,
   socialWorkerSummaryList,
@@ -47,8 +48,8 @@ export const enContent = {
   },
   sectionTitles: {
     applicationDetails: 'Application details',
-    adoptionagencyOrLA: 'Adoption agency or local authority details',
-    additionalAoptionagencyOrLA: 'Additional adoption agency or local authority details',
+    adoptionagencyOrLA: 'Local authority details',
+    additionalAoptionagencyOrLA: 'Adoption agency or local authority details',
     socialWorkerDetails: "Child's social worker details",
     applicantDetails: "Applicant's details",
     firstApplicantDetails: "First applicant's details",
@@ -70,7 +71,7 @@ export const enContent = {
     emailAddress: 'Email address',
     nameOfContact: 'Name of contact',
     emailOfContact: 'Email address of contact',
-    teamEmailAddress: 'Team email address',
+    childLocalAuthority: "Child's local authority",
     additionalAdoptionAgency: 'Additional adoption agency',
     fullName: 'Full name',
     previousNames: 'Previous names',
@@ -111,8 +112,8 @@ const en = (content: CommonContent) => {
     language: content.language,
     sections: [
       applicationSummaryList(enContent, userCase),
+      localAuthoritySummaryList(enContent, userCase),
       adoptionAgencySummaryList(enContent, userCase),
-      adoptionAgencySummaryList(enContent, userCase, 1),
       socialWorkerSummaryList(enContent, userCase),
       applicantSummaryList(enContent, userCase, FieldPrefix.APPLICANT1),
       applicantSummaryList(enContent, userCase, FieldPrefix.APPLICANT2),
@@ -157,8 +158,8 @@ const cyContent: typeof enContent = {
   },
   sectionTitles: {
     applicationDetails: 'Manylion y cais',
-    adoptionagencyOrLA: 'Manylion yr asiantaeth fabwysiadu neu’r awdurdod lleol',
-    additionalAoptionagencyOrLA: 'Manylion ychwanegol yr asiantaeth fabwysiadu neu’r awdurdod lleol',
+    adoptionagencyOrLA: 'Local authority details (in welsh)',
+    additionalAoptionagencyOrLA: 'Adoption agency or local authority details (in welsh)',
     socialWorkerDetails: 'Manylion gweithiwr cymdeithasol y plentyn',
     applicantDetails: 'Manylion y ceisydd',
     firstApplicantDetails: 'Manylion y ceisydd cyntaf',
@@ -180,7 +181,7 @@ const cyContent: typeof enContent = {
     emailAddress: 'Cyfeiriad e-bost',
     nameOfContact: 'Enw’r unigolyn cyswllt',
     emailOfContact: 'Cyfeiriad e-bost yr unigolyn cyswllt',
-    teamEmailAddress: 'Cyfeiriad e-bost y tîm',
+    childLocalAuthority: "Child's local authority (in welsh)",
     additionalAdoptionAgency: 'Asiantaeth fabwysiadu ychwanegol',
     fullName: 'Enw llawn',
     previousNames: 'Enwau blaenorol',
@@ -222,8 +223,8 @@ const cy: typeof en = (content: CommonContent) => {
     language: content.language,
     sections: [
       applicationSummaryList(cyContent, userCase),
+      localAuthoritySummaryList(cyContent, userCase),
       adoptionAgencySummaryList(cyContent, userCase),
-      adoptionAgencySummaryList(cyContent, userCase, 1),
       socialWorkerSummaryList(cyContent, userCase),
       applicantSummaryList(cyContent, userCase, FieldPrefix.APPLICANT1),
       applicantSummaryList(cyContent, userCase, FieldPrefix.APPLICANT2),
