+++ src/main/steps/task-list/utils.test.ts
@@ -778,100 +778,6 @@ describe('utils', () => {
     });
   });
 
-  // describe('getAdoptionAgencyUrl()', () => {
-  //   // test('Should return adoption agency add url when adoption-agency count 0', async () => {
-  //   //   expect(getAdoptionAgencyUrl(userCase)).toBe('/children/adoption-agency?add=1609459200000');
-  //   // });
-
-  // test('Should return adoption agency add url when adoption-agency count: 2 and hasAnotherAdopAgencyOrLA: NO', async () => {
-  //   expect(
-  //     getAdoptionAgencyUrl({
-  //       id: '123',
-  //       state: State.Draft,
-  //       connections: [],
-  //       documentsGenerated: [],
-  //       applicationFeeOrderSummary: { PaymentReference: '', Fees: [], PaymentTotal: '0' },
-  //       adopAgencyOrLAs: [
-  //         {
-  //           adopAgencyOrLaId: '1',
-  //           adopAgencyOrLaName: 'a',
-  //           adopAgencyOrLaContactName: 'a',
-  //           adopAgencyOrLaPhoneNumber: '09876543210',
-  //           adopAgencyOrLaContactEmail: 'a@a.a',
-  //         },
-  //         {
-  //           adopAgencyOrLaId: '2',
-  //           adopAgencyOrLaName: 'a',
-  //           adopAgencyOrLaContactName: 'a',
-  //           adopAgencyOrLaPhoneNumber: '09876543210',
-  //           adopAgencyOrLaContactEmail: 'a@a.a',
-  //         },
-  //       ],
-  //       hasAnotherAdopAgencyOrLA: YesOrNo.NO,
-  //       socialWorkerName: 'John',
-  //       socialWorkerPhoneNumber: '09876543210',
-  //       socialWorkerEmail: 'abc@john.com',
-  //     })
-  //   ).toBe('/children/adoption-agency');
-  // });
-
-  // test('Should return adoption agency add url when adoption-agency count: 2 and hasAnotherAdopAgencyOrLA: Yes', async () => {
-  //   expect(
-  //     getAdoptionAgencyUrl({
-  //       id: '123',
-  //       state: State.Draft,
-  //       connections: [],
-  //       documentsGenerated: [],
-  //       applicationFeeOrderSummary: { PaymentReference: '', Fees: [], PaymentTotal: '0' },
-  //       adopAgencyOrLAs: [
-  //         {
-  //           adopAgencyOrLaId: '1',
-  //           adopAgencyOrLaName: 'a',
-  //           adopAgencyOrLaContactName: 'a',
-  //           adopAgencyOrLaPhoneNumber: '09876543210',
-  //           adopAgencyOrLaContactEmail: 'a@a.a',
-  //         },
-  //         {
-  //           adopAgencyOrLaId: '2',
-  //           adopAgencyOrLaName: 'a',
-  //           adopAgencyOrLaContactName: 'a',
-  //           adopAgencyOrLaPhoneNumber: '09876543210',
-  //           adopAgencyOrLaContactEmail: 'a@a.a',
-  //         },
-  //       ],
-  //       hasAnotherAdopAgencyOrLA: YesOrNo.YES,
-  //       socialWorkerName: 'John',
-  //       socialWorkerPhoneNumber: '09876543210',
-  //       socialWorkerEmail: 'abc@john.com',
-  //     })
-  //   ).toBe('/children/adoption-agency?change=1');
-  // });
-
-  // test('Should return adoption agency add url when adoption-agency count: 1, hasAnotherAdopAgencyOrLA: NO and phone number not available', async () => {
-  //   expect(
-  //     getAdoptionAgencyUrl({
-  //       id: '123',
-  //       state: State.Draft,
-  //       connections: [],
-  //       documentsGenerated: [],
-  //       applicationFeeOrderSummary: { PaymentReference: '', Fees: [], PaymentTotal: '0' },
-  //       adopAgencyOrLAs: [
-  //         {
-  //           adopAgencyOrLaId: '1',
-  //           adopAgencyOrLaName: 'a',
-  //           adopAgencyOrLaContactName: 'a',
-  //           adopAgencyOrLaContactEmail: 'a@a.a',
-  //         },
-  //       ],
-  //       hasAnotherAdopAgencyOrLA: YesOrNo.NO,
-  //       socialWorkerName: 'John',
-  //       socialWorkerPhoneNumber: '09876543210',
-  //       socialWorkerEmail: 'abc@john.com',
-  //     })
-  //   ).toBe('/children/adoption-agency?change=1');
-  // });
-  // });
-
   describe('getSiblingPlacementOrderStatus', () => {
     test.each([
       { data: {}, expected: 'NOT_STARTED' },
